+++ Changelog.md
@@ -1,6 +1,7 @@
 # version 0.87.2
-- bug removed : master private/public keys when None and more
+- bug removed: master private/public keys when None and more
 - few clippy warnings resolved
+- learned (again) about rust filtering
 
 # version 0.87.1
 - bug removed: panic when public/private header is None